defconfig: oneplus5: Properly display version for inline builds
Since this kernel has to be built inline with a ROM, we need to define
the kernel name and enable the automatic version string based on the git
revision in order to display the custom kernel name and version. This
was being previously done by my build script for standalone builds.

Signed-off-by: Michele Beccalossi <m.beccalossi003@studenti.unibs.it>